* master: (5045 commits)
  svn update
  update addon scripts
  patch [#28218] During-render callback functionality from Jesse Kaukonen (gekko)
  COLLADA: Take parent bone length and direction instead of using bone pointing up with length 1. Looks much nicer and less confusing on larger armatures now.
  BGE: Adding two new functions to bge.render to allow users to change the anisotropic filtering level used by textures:   * setAnisotropicFiltering(level)   * getAnisotropicFiltering()
  typo fix: end of lines ;; --> ;
  fix [#28430] Image with Stampinfo does not get saved correctly with alpha
  OSX: additionally choice to link against python-framework again
  fix for error in my recent change to image save. - relative path wasn't being made absolute. - saving renders was always defaulting to multilayer exr, now use the output format set.
  catch exception and report an error when failing to write exr files - was crashing with debug builds.
  SVN maintenance.
  When creating a multisample window for Win32, the first context created, which checks for multisample support, was never being deleted.
  Check for potential crasher.
  Fix #28427: smooth faces flash momentarily when extruded using "extrude and move on normals" tool (E key)
  Fix for [#28425] when user prefs -> editing -> align to == "view" newly inserted objects do not show the applied rotation in the tools panel
  2.5 Game UI: *Fix for clutter after pepper merge, 3 booleans in one row, is 1 too much here ;-)
  Enable libsndfile by default on win32 too.
  Fix for my last commit.
  minor edits, pep8 - also correct float -> double promotion for blf.
  2.5 UI Files: * Code cleanup after Pepper merge.
  ...

Conflicts:
	source/blender/editors/space_view3d/drawobject.c
	source/blender/makesdna/DNA_object_types.h

+++ source/blender/blenlib/intern/jitter.c
@@ -1,4 +1,4 @@
-/**
+/*
  * Jitter offset table
  *
  * $Id$
@@ -29,11 +29,17 @@
  * ***** END GPL LICENSE BLOCK *****
  */
 
+/** \file blender/blenlib/intern/jitter.c
+ *  \ingroup bli
+ */
+
+
 #include <math.h>
 #include <string.h>
 #include "MEM_guardedalloc.h"
 
 #include "BLI_rand.h"
+#include "BLI_jitter.h"
 
 
 void BLI_jitterate1(float *jit1, float *jit2, int num, float rad1)
@@ -47,10 +53,10 @@ void BLI_jitterate1(float *jit1, float *jit2, int num, float rad1)
 		y = jit1[i+1];
 		for (j = 2*num-2; j>=0 ; j-=2) {
 			if (i != j){
-				vecx = jit1[j] - x - 1.0;
-				vecy = jit1[j+1] - y - 1.0;
+				vecx = jit1[j] - x - 1.0f;
+				vecy = jit1[j+1] - y - 1.0f;
 				for (k = 3; k>0 ; k--){
-					if( fabs(vecx)<rad1 && fabs(vecy)<rad1) {
+					if( fabsf(vecx)<rad1 && fabsf(vecy)<rad1) {
 						len=  sqrt(vecx*vecx + vecy*vecy);
 						if(len>0 && len<rad1) {
 							len= len/rad1;
@@ -58,9 +64,9 @@ void BLI_jitterate1(float *jit1, float *jit2, int num, float rad1)
 							dvecy += vecy/len;
 						}
 					}
-					vecx += 1.0;
+					vecx += 1.0f;
 
-					if( fabs(vecx)<rad1 && fabs(vecy)<rad1) {
+					if( fabsf(vecx)<rad1 && fabsf(vecy)<rad1) {
 						len=  sqrt(vecx*vecx + vecy*vecy);
 						if(len>0 && len<rad1) {
 							len= len/rad1;
@@ -68,9 +74,9 @@ void BLI_jitterate1(float *jit1, float *jit2, int num, float rad1)
 							dvecy += vecy/len;
 						}
 					}
-					vecx += 1.0;
+					vecx += 1.0f;
 
-					if( fabs(vecx)<rad1 && fabs(vecy)<rad1) {
+					if( fabsf(vecx)<rad1 && fabsf(vecy)<rad1) {
 						len=  sqrt(vecx*vecx + vecy*vecy);
 						if(len>0 && len<rad1) {
 							len= len/rad1;
@@ -78,16 +84,16 @@ void BLI_jitterate1(float *jit1, float *jit2, int num, float rad1)
 							dvecy += vecy/len;
 						}
 					}
-					vecx -= 2.0;
-					vecy += 1.0;
+					vecx -= 2.0f;
+					vecy += 1.0f;
 				}
 			}
 		}
 
-		x -= dvecx/18.0 ;
-		y -= dvecy/18.0;
-		x -= floor(x) ;
-		y -= floor(y);
+		x -= dvecx/18.0f;
+		y -= dvecy/18.0f;
+		x -= floorf(x) ;
+		y -= floorf(y);
 		jit2[i] = x;
 		jit2[i+1] = y;
 	}
@@ -105,28 +111,28 @@ void BLI_jitterate2(float *jit1, float *jit2, int num, float rad2)
 		y = jit1[i+1];
 		for (j =2*num -2; j>= 0 ; j-=2){
 			if (i != j){
-				vecx = jit1[j] - x - 1.0;
-				vecy = jit1[j+1] - y - 1.0;
+				vecx = jit1[j] - x - 1.0f;
+				vecy = jit1[j+1] - y - 1.0f;
 
-				if( fabs(vecx)<rad2) dvecx+= vecx*rad2;
-				vecx += 1.0;
-				if( fabs(vecx)<rad2) dvecx+= vecx*rad2;
-				vecx += 1.0;
-				if( fabs(vecx)<rad2) dvecx+= vecx*rad2;
+				if( fabsf(vecx)<rad2) dvecx+= vecx*rad2;
+				vecx += 1.0f;
+				if( fabsf(vecx)<rad2) dvecx+= vecx*rad2;
+				vecx += 1.0f;
+				if( fabsf(vecx)<rad2) dvecx+= vecx*rad2;
 
-				if( fabs(vecy)<rad2) dvecy+= vecy*rad2;
-				vecy += 1.0;
-				if( fabs(vecy)<rad2) dvecy+= vecy*rad2;
-				vecy += 1.0;
-				if( fabs(vecy)<rad2) dvecy+= vecy*rad2;
+				if( fabsf(vecy)<rad2) dvecy+= vecy*rad2;
+				vecy += 1.0f;
+				if( fabsf(vecy)<rad2) dvecy+= vecy*rad2;
+				vecy += 1.0f;
+				if( fabsf(vecy)<rad2) dvecy+= vecy*rad2;
 
 			}
 		}
 
-		x -= dvecx/2 ;
-		y -= dvecy/2;
-		x -= floor(x) ;
-		y -= floor(y);
+		x -= dvecx/2.0f;
+		y -= dvecy/2.0f;
+		x -= floorf(x) ;
+		y -= floorf(y);
 		jit2[i] = x;
 		jit2[i+1] = y;
 	}
@@ -142,17 +148,17 @@ void BLI_initjit(float *jitarr, int num)
 	if(num==0) return;
 
 	jit2= MEM_mallocN(12 + 2*sizeof(float)*num, "initjit");
-	rad1=  1.0/sqrt((float)num);
-	rad2= 1.0/((float)num);
-	rad3= sqrt((float)num)/((float)num);
+	rad1= 1.0f/sqrtf((float)num);
+	rad2= 1.0f/((float)num);
+	rad3= sqrtf((float)num)/((float)num);
 
 	BLI_srand(31415926 + num);
 	x= 0;
 	for(i=0; i<2*num; i+=2) {
-		jitarr[i]= x+ rad1*(0.5-BLI_drand());
-		jitarr[i+1]= ((float)i/2)/num +rad1*(0.5-BLI_drand());
+		jitarr[i]= x+ rad1*(float)(0.5-BLI_drand());
+		jitarr[i+1]= ((float)i/2)/num +rad1*(float)(0.5-BLI_drand());
 		x+= rad3;
-		x -= floor(x);
+		x -= floorf(x);
 	}
 
 	for (i=0 ; i<24 ; i++) {
@@ -165,8 +171,8 @@ void BLI_initjit(float *jitarr, int num)
 	
 	/* finally, move jittertab to be centered around (0,0) */
 	for(i=0; i<2*num; i+=2) {
-		jitarr[i] -= 0.5;
-		jitarr[i+1] -= 0.5;
+		jitarr[i] -= 0.5f;
+		jitarr[i+1] -= 0.5f;
 	}
 	
 }
